Exercise exact bindings in EditMode
Share the existing direct-query oracle through runtime-capable test support and expose the same twelve cases in an Editor fixture. Register its CommonTestBase inheritance in the lifecycle linter and require mode-specific execution evidence in the fixture-runner skill. Native EditMode detects the original refusal in all twelve cases, then passes all twelve with the production file restored byte-identically. The full Runtime fixture passes twenty-five cases. Host Runtime and Editor variants, assembly-reference probes, 120 linter regressions, documentation counts, and changed-file preflight pass.
