+++ Tests/Core/RelationalExactBindingTestBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 00fc08f1b013da4b1544dad04310c5db
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
