+++ Tests/Editor/Attributes/RelationalExactBindingTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cf74819c6df5defadc095e2626a8d75b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
